Implemented size_t for sizes and indices

+++ Ultrasonic_Wrapper.cpp
@@ -4,7 +4,7 @@
 Ultrasonic_Wrapper::Ultrasonic_Wrapper(unsigned int trigPin, 
                                        unsigned int* echoPins,
                                        unsigned int* burstFrequencies, 
-                                       unsigned int sensorNum) :
+                                       size_t sensorNum) :
                                        _trigPin(trigPin),
                                        _sensorNum(sensorNum)
 {
@@ -14,7 +14,7 @@ Ultrasonic_Wrapper::Ultrasonic_Wrapper(unsigned int trigPin,
 Ultrasonic_Wrapper::Ultrasonic_Wrapper(unsigned int trigPin,
                                        unsigned int* echoPins,
                                        unsigned int burstFrequency,
-                                       unsigned int sensorNum) :
+                                       size_t sensorNum) :
                                        _trigPin(trigPin),
                                        _sensorNum(sensorNum)
 {
@@ -32,7 +32,7 @@ Ultrasonic_Wrapper::Ultrasonic_Wrapper(unsigned int trigPin,
 
 Ultrasonic_Wrapper::~Ultrasonic_Wrapper()
 {
-    for (unsigned int i = 0; i < _sensorNum; i++)
+    for (size_t i = 0; i < _sensorNum; i++)
     {
         delete _ultrasonicsPtr[i];
     }
@@ -41,7 +41,7 @@ Ultrasonic_Wrapper::~Ultrasonic_Wrapper()
 
 void Ultrasonic_Wrapper::begin(void (*externalEchoPinISRs[])())
 {
-    for (unsigned int i = 0; i < _sensorNum; i++)
+    for (size_t i = 0; i < _sensorNum; i++)
     {
         _ultrasonicsPtr[i]->begin(externalEchoPinISRs[i]);
     }
@@ -49,7 +49,7 @@ void Ultrasonic_Wrapper::begin(void (*externalEchoPinISRs[])())
 
 void Ultrasonic_Wrapper::begin(void (*externalEchoPinISR)())
 {
-    for (unsigned int i = 0; i < _sensorNum; i++)
+    for (size_t i = 0; i < _sensorNum; i++)
     {
         _ultrasonicsPtr[i]->begin(externalEchoPinISR);
     }
@@ -57,18 +57,18 @@ void Ultrasonic_Wrapper::begin(void (*externalEchoPinISR)())
 
 void Ultrasonic_Wrapper::update()
 {
-    for (unsigned int i = 0; i < _sensorNum; i++)
+    for (size_t i = 0; i < _sensorNum; i++)
     {
         _ultrasonicsPtr[i]->update();
     }
 }
 
-unsigned int Ultrasonic_Wrapper::getDistance(unsigned int index /*= 0*/) const
+unsigned int Ultrasonic_Wrapper::getDistance(size_t index /*= 0*/) const
 {
     return _ultrasonicsPtr[index]->getDistance();
 }
 
-unsigned int Ultrasonic_Wrapper::getEchoPin(unsigned int index /*= 0*/) const
+unsigned int Ultrasonic_Wrapper::getEchoPin(size_t index /*= 0*/) const
 {
     return _ultrasonicsPtr[index]->getEchoPin();
 }
@@ -78,7 +78,7 @@ unsigned int Ultrasonic_Wrapper::getTrigPin() const
     return _trigPin;
 }
 
-void Ultrasonic_Wrapper::echoPinISR(unsigned int index /*= 0*/)
+void Ultrasonic_Wrapper::echoPinISR(size_t index /*= 0*/)
 {
     _ultrasonicsPtr[index]->echoPinISR();
 }
@@ -86,7 +86,7 @@ void Ultrasonic_Wrapper::echoPinISR(unsigned int index /*= 0*/)
 void Ultrasonic_Wrapper::_init(unsigned int *echoPins, unsigned int *burstFrequencies)
 {
     _ultrasonicsPtr = new Ultrasonic *[_sensorNum];
-    for (unsigned int i = 0; i < _sensorNum; i++)
+    for (size_t i = 0; i < _sensorNum; i++)
     {
         _ultrasonicsPtr[i] = new Ultrasonic(_trigPin, *echoPins++, *burstFrequencies++);
     }
